# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Slicer API\n基于Kubernetes资源的切片管理系统API\n包括 切片管理 监控管理 性能保证 等功能",
        "title": "Slicer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:30001",
    "basePath": "/",
    "paths": {
...
        "/play": {
            "post": {
                "description": "接收Play对象并创建新资源，同时部署到Kubernetes集群",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Play"
                ],
                "summary": "创建Play资源",
                "parameters": [
                    {
                        "description": "Play对象",
                        "name": "play",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Play"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功返回Play对象",
                        "schema": {
                            "$ref": "#/definitions/model.Play"
                        }
                    },
                    "400": {
                        "description": "请求解码失败/参数非法/Slice不存在/Play已存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "关联Slice不存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "存储失败/部署失败/响应编码失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/play/{playId}": {
            "get": {
                "description": "根据Play ID获取资源详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Play"
                ],
                "summary": "获取单个Play",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Play ID",
                        "name": "playId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "$ref": "#/definitions/model.Play"
                        }
                    },
                    "400": {
                        "description": "缺少Play ID",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Play不存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "获取失败/响应编码失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "根据Play ID更新资源并重新部署",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Play"
                ],
                "summary": "更新Play资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Play ID",
                        "name": "playId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新后的Play对象",
                        "name": "play",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Play"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功返回对象",
                        "schema": {
                            "$ref": "#/definitions/model.Play"
                        }
                    },
                    "400": {
                        "description": "缺少Play ID/请求解码失败/参数非法",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Play不存在",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "更新失败/部署失败/响应编码失败",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Play": {
            "type": "object"
        },
...
    }
}